1.840u.c驱动代码部分逻辑使用变量与定义不一致导致编不过问题，同步差异编译可以通过。 2.在qm和mtk上增加76xx代码，在qm上增加840u代码。 3.由于76xx升级固件len值过大超过u16，所有将updta_fw函数的len传参统一从u16改为u32 4.840u和76xx作为新做代码，初始化流程在拉复位后增加先在main中读取信息的judge_moudle流程，通过宏JUDGE_MODULE_EN来控制打开。 5.esd流程由于新的840和76xx返回值的流程不一致，所有增加宏ESD_READ_TIME_EN来控制打开读取时间返回值，不开时默认返回新sdk中设置的固定1字节返回值的读取状态，0;读取正常，1:ic正常模式不匹配，2：ic异常。

+++ hyn_driver_release_mtk/hyn_chips/hyn_cst3240.c
@@ -397,7 +397,7 @@ static int cst3240_updata_judge(u8 *p_fw, u16 len)
     return 0;
 }
 
-static int cst3240_updata_fw(u8 *bin_addr, u16 len)
+static int cst3240_updata_fw(u8 *bin_addr, u32 len)
 {
     int i,ret=-1, times = 0,retry;
 	u8 i2c_buf[512+2];
