+++ hyn_driver_release_qm/hyn_chips/hyn_cst226se.c
@@ -333,7 +333,7 @@ static int cst226se_updata_judge(u8 *p_fw, u16 len)
     return 0;
 }
 
-static int cst226se_updata_fw(u8 *bin_addr, u16 len)
+static int cst226se_updata_fw(u8 *bin_addr, u32 len)
 {
     #define CHECKSUM_OFFECT  (7680-4)
     int i,ret, retry = 4;
